+++ testing/upload_validation.ts
@@ -382,7 +382,7 @@ const defaultAuthenticationValidators: Record<AuthenticationType, z.ZodTypeAny> 
     endpointKey: z.string().optional(),
     tokenQueryParam: z.string().optional(),
     useProofKeyForCodeExchange: z.boolean().optional(),
-    suppressScopeParam: z.boolean().optional(),
+    scopeParamName: z.string().optional(),
     nestedResponseKey: z.string().optional(),
     ...baseAuthenticationValidators,
   }),